fix install by not importing requirements during setup
