+++ .coveragerc
@@ -3,7 +3,6 @@ branch = True
 omit =
     */MedicalImageAnonymizer/__init__.py
     */MedicalImageAnonymizer/__version__.py
-    */MedicalImageAnonymizer/build.py
     */MedicalImageAnonymizer/GUI/__init__.py
 [report]
 # Regexes for lines to exclude from consideration
