+++ SOURCES.txt
@@ -1,7 +1,6 @@
 setup.py
 MedicalImageAnonymizer/__init__.py
 MedicalImageAnonymizer/__version__.py
-MedicalImageAnonymizer/build.py
 MedicalImageAnonymizer/Anonymizer.py
 MedicalImageAnonymizer/DICOM_anonymizer.py
 MedicalImageAnonymizer/Nifti_anonymizer.py
